consolidated role assignments to 1 file/resource
